requirements.txt
fastapi
uvicorn
llama-index
PyPDF2
google-api-python-client
google-auth